Refactor AutomatonController and update related files
Significantly modified the AutomatonController to include a logger and change the `SimulateDfa` method to use `DfaViewModel`. Removed `SimulateNfa` and `MinimizeDfa` methods, along with the `AutomatonDto` class.

Commented out the `SimulateDfa` method in HomeController, disabling its functionality. Updated the form action in Index.cshtml to point to the new `SimulateDfa` method in AutomatonController, ensuring proper front-end integration.